+++ controllers/usuario.js
@@ -1,22 +1,6 @@
 const db = require('../config/db');
 const argon = require('argon2'); // função de hash
-const crypto = require('crypto');
-const TAMANHO_SALT = 64;
-
-function gerarSalt() {
-    return crypto.randomBytes(Math.ceil(TAMANHO_SALT / 2))
-        .toString('hex')
-        .slice(0, TAMANHO_SALT);
-}
-
-async function verificarHash(hash, senha, salt) {
-    console.log(hash, senha, salt)
-    return ;
-}
-
-async function hashSenha(senha, salt) {
-    return argon.hash(senha + salt)
-}
+const { getUsuario, gerarSalt } = require('./utils');
 
 async function getUsuarioCreate(req, res) {
     res.render('usuario/usuarioCreate', { layout: 'noMenu.handlebars' })
@@ -24,7 +8,7 @@ async function getUsuarioCreate(req, res) {
 
 async function postUsuarioCreate(req, res) {
     const novoSalt = gerarSalt();
-    const novoHash = await hashSenha(req.body.senha, novoSalt);
+    const novoHash = await argon.hash(req.body.senha + novoSalt);
 
     db.Usuario.create({
         nome: req.body.nome,
@@ -65,27 +49,28 @@ async function getUsuarioSelfEdit(req, res) {
             id: req.session.id,
         }
     }).then((usuario) => {
-        res.render('usuario/usuarioEdit', { usuario: usuario.toJSON() });
+        res.render('usuario/usuarioEdit', { 
+            usuario: usuario.toJSON(),
+            layout: 'noMenu.handlebars'
+        });
     }).catch(err => {
         res.render('erros/usuarioNaoAchado');
     });
 }
 
 async function postUsuarioEdit(req, res) {
-    const novoSalt = gerarSalt();
-    const novoHash = await hashSenha(req.body.senha, novoSalt);
-
+    const eAdmin = req.body.eAdmin ? true : false;
+    const eColaborador = req.body.eColaborador ? true : false;
+    
     db.Usuario.findOne({
         where: { id: req.body.id }
     }).then(usuario => {
         usuario.set({
             nome: req.body.nome,
             sobrenome: req.body.sobrenome,
-            senha: novoHash,
-            salt: novoSalt,
             email: req.body.email,
-            eAtivo: req.body.eAtivo,
-            eColaborador: req.body.eColaborador,
+            eAdmin: eAdmin,
+            eColaborador: eColaborador,
         });
         usuario.save();
     }).catch(error => {
@@ -98,11 +83,13 @@ async function postUsuarioEdit(req, res) {
 }
 
 async function getUsuarioDelete(req, res) {
-    db.Usuario.destroy({
-        where: {
-            id: req.params.id,
-        }
-    });
+    if (req.session.eAdmin) {
+        db.Usuario.destroy({
+            where: {
+                id: req.params.id,
+            }
+        });
+    }
 }
 
 async function getUsuarioLogin(req, res) {
@@ -115,19 +102,20 @@ async function postUsuarioLogin(req, res) {
     db.Usuario.findOne({
         where: { email: email }
     }).then(usuario => {
-        console.log(senha + email);
-        console.log(usuario);
-        argon.verify(usuario.senha, senha + usuario.salt)
-            .then(deu_certo => {
+        if (usuario) {
+            argon.verify(usuario.senha, senha + usuario.salt).then(deu_certo => {
                 if (deu_certo) {
                     req.session.id_usuario = usuario.id;
                     req.session.nome = usuario.nome;
                     req.session.eAdmin = usuario.eAdmin;
                     res.redirect('/home')
                 } else {
-                    res.render('erros/usuarioNaoAchado')
+                    throw new Error("Senha inválida")
                 }
             });
+        } else {
+            throw new Error("E-mail não cadastrado.")
+        }
     }).catch(error => {
         res.render('erros/usuarioNaoAchado', {error: error})
     });
@@ -139,9 +127,20 @@ async function getUsuarioLogout(req, res) {
 }
 
 async function getUsuarioList(req, res) {
-    res.json({
-        msg: "Em construção"
-    })
+    db.Usuario.findAll().then(usuarios => {
+        var lista = usuarios.map(usuarios => usuarios.toJSON())
+        res.render('usuario/usuarioList', {
+            usuario: getUsuario(req),
+            listaUsuarios: lista,
+            layout: 'main.handlebars' 
+        });
+    }).catch(error => {
+        res.render('usuario/usuarioList', {
+            error: error,
+            usuario: getUsuario(req),
+            layout: 'main.handlebars',
+        });
+    });
 }
 
 module.exports = {
